Garbage: fixed indexOutOfBound error for Jack
Couldn't really figure out why it wasn't working, but I made it so that the process of picking a card to place the jack was the same as if a player was doing it, meaning instead of the bot removing and handing a card, it would return a number of where it wanted the card.
Now I need to fix the win condition and fix if the last card left is a jack to leave it there instead of trying to find a place for it.